api/requirements.txt
fastapi
uvicorn
deepface
pillow
tensorflow==2.13.0
h5py